Module psas_telemetry::messages

Modules

adis
Module for the ADIS16405 IMU telemetry messages.

bmp1
Module for the BMP180 Pressure sensor telemetry messages.

fcfh
Module for the FC Health telemetry messages.

rnhh
Module for the RNH Health telemetry messages.

rnhp
Module for the Rocket Net Hub Power telemetry messages.

rnhu
Module for the Umbilical Detect telemetry messages.

roll
Module for the Roll Servo Commands telemetry messages.

seqe
Module for the Sequence Number Error telemetry messages.

v8a8
Module for the Venus8 Navigation Message telemetry messages.

Structs

Header
A message header

Functions

read_header
Read from a buffer assuming were at the boundary of a message. Extracts the header

read_seqn_body

to_csv
Attempt to read Message body based on it's fourcc and convert to a CSV line of scaled, unit-converted values

to_values
Attempt to read Message body based on it's fourcc and convert to a map of scaled, unit-converted values