Struct psas_telemetry::messages::adis::ADIS

pub struct ADIS {
    pub vcc: f64,
    pub gyro_x: f64,
    pub gyro_y: f64,
    pub gyro_z: f64,
    pub acc_x: f64,
    pub acc_y: f64,
    pub acc_z: f64,
    pub magn_x: f64,
    pub magn_y: f64,
    pub magn_z: f64,
    pub temp: f64,
    pub aux_adc: f64,
}

Raw data from the Analog Devices ADIS16405 9DOF IMU

Fields
vcc: f64

Bus voltage [Volt]

gyro_x: f64

X axis value from rate-gyroscope [deg/s]

gyro_y: f64

Y axis value from rate-gyroscope [deg/s]

gyro_z: f64

Z axis value from rate-gyroscope [deg/s]

acc_x: f64

X axis value from accelerometer [g]

acc_y: f64

Y axis value from accelerometer [g]

acc_z: f64

Z axis value from accelerometer [g]

magn_x: f64

X axis value from magnetometer [tesla]

magn_y: f64

Y axis value from magnetometer [tesla]

magn_z: f64

Z axis value from magnetometer [tesla]

temp: f64

Unit temperature [°C]

aux_adc: f64

Aux ADC value [volt]

Methods

impl ADIS

pub fn from_raw(raw: ADIS_raw) -> ADIS

Convert a ADIS_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})