Struct psas_telemetry::messages::bmp1::BMP1

pub struct BMP1 {
    pub pressure: f64,
    pub temperature: f64,
}

Raw data from BMP180

Fields
pressure: f64

Pressure data []

temperature: f64

Temperature data []

Methods

impl BMP1

pub fn from_raw(raw: BMP1_raw) -> BMP1

Convert a BMP1_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})