Struct psas_telemetry::messages::fcfh::FCFH

pub struct FCFH {
    pub cpu_user: f64,
    pub cpu_system: f64,
    pub cpu_nice: f64,
    pub cpu_iowait: f64,
    pub cpu_irq: f64,
    pub cpu_softirq: f64,
    pub ram_used: f64,
    pub ram_buffer: f64,
    pub ram_cached: f64,
    pub pid: f64,
    pub disk_used: f64,
    pub disk_buffer: f64,
    pub disk_cached: f64,
    pub io_lo_bytes_sent: f64,
    pub io_lo_bytes_recv: f64,
    pub io_lo_packets_sent: f64,
    pub io_lo_packets_recv: f64,
    pub io_eth0_bytes_sent: f64,
    pub io_eth0_bytes_recv: f64,
    pub io_eth0_packets_sent: f64,
    pub io_eth0_packets_recv: f64,
    pub io_wlan0_bytes_sent: f64,
    pub io_wlan0_bytes_recv: f64,
    pub io_wlan0_packets_sent: f64,
    pub io_wlan0_packets_recv: f64,
    pub core_temp: f64,
}

Main flight computer software health

Fields
cpu_user: f64

CPU use in user space [%]

cpu_system: f64

CPU use in kernel space [%]

cpu_nice: f64

CPU use in running users' processes that have been 'niced' [%]

cpu_iowait: f64

CPU use spent waiting for I/O to complete [%]

cpu_irq: f64

CPU use spent for servicing hardware interrupts [%]

cpu_softirq: f64

CPU use spent for servicing software interrupts [%]

ram_used: f64

Ram currently in use [Bytes]

ram_buffer: f64

Ram currently in use, cache for things like file system metadata [Bytes]

ram_cached: f64

Ram currently in use, cache for various things [Bytes]

pid: f64

Number of PIDs currently running [#]

disk_used: f64

Amount of disk space currently used [Bytes]

disk_buffer: f64

Amount of disk space currently buffered [Bytes]

disk_cached: f64

Amount of disk space currently cached [Bytes]

io_lo_bytes_sent: f64

Bytes sent on the loopback interface (localhost) [Bytes]

io_lo_bytes_recv: f64

Bytes recieved on the loopback interface (localhost) [Bytes]

io_lo_packets_sent: f64

Packets sent on the loopback interface (localhost) [#]

io_lo_packets_recv: f64

Packets recieved on the loopback interface (localhost) [#]

io_eth0_bytes_sent: f64

Bytes sent on the eth0 interface [Bytes]

io_eth0_bytes_recv: f64

Bytes recieved on the eth0 interface [Bytes]

io_eth0_packets_sent: f64

Packets sent on the eth0 interface [#]

io_eth0_packets_recv: f64

Packets recieved on the eth0 interface [#]

io_wlan0_bytes_sent: f64

Bytes sent on the wlan0 interface [Bytes]

io_wlan0_bytes_recv: f64

Bytes recieved on the wlan0 interface [Bytes]

io_wlan0_packets_sent: f64

Packets sent on the wlan0 interface [#]

io_wlan0_packets_recv: f64

Packets recieved on the wlan0 interface [#]

core_temp: f64

Temperature of the processor [°C]

Methods

impl FCFH

pub fn from_raw(raw: FCFH_raw) -> FCFH

Convert a FCFH_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})