Struct psas_telemetry::messages::rnhh::RNHH

pub struct RNHH {
    pub temperature: f64,
    pub ts1temperature: f64,
    pub ts2temperature: f64,
    pub temprange: f64,
    pub voltage: f64,
    pub current: f64,
    pub averagecurrent: f64,
    pub cellvoltage1: f64,
    pub cellvoltage2: f64,
    pub cellvoltage3: f64,
    pub cellvoltage4: f64,
    pub packvoltage: f64,
    pub averagevoltage: f64,
}

Rocket Net Hub hardware Health

Fields
temperature: f64

Temperature of the main PCB [Kelvin]

ts1temperature: f64

Temperature of battery sensor 1 [°C]

ts2temperature: f64

Temperature of battery sensor 2 [°C]

temprange: f64

[]

voltage: f64

System voltage [Volts]

current: f64

System current [Amps]

averagecurrent: f64

System current (average) [Amps]

cellvoltage1: f64

Battery cell #1 volts [Volt]

cellvoltage2: f64

Battery cell #2 volts [Volt]

cellvoltage3: f64

Battery cell #3 volts [Volt]

cellvoltage4: f64

Battery cell #4 volts [Volt]

packvoltage: f64

Battery pack voltage [Volt]

averagevoltage: f64

Battery pack voltage (average) [Volt]

Methods

impl RNHH

pub fn from_raw(raw: RNHH_raw) -> RNHH

Convert a RNHH_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})