Struct psas_telemetry::messages::rnhp::RNHP

pub struct RNHP {
    pub port1: f64,
    pub port2: f64,
    pub port3: f64,
    pub port4: f64,
    pub umbilical: f64,
    pub port6: f64,
    pub port7: f64,
    pub port8: f64,
}

Power stats from the RNH

Fields
port1: f64

Port 1 power [Amp]

port2: f64

Port 2 power [Amp]

port3: f64

Port 3 power [Amp]

port4: f64

Port 4 power [Amp]

umbilical: f64

Umbilical power, power entering the rocket from shore power [Amp]

port6: f64

Port 6 power [Amp]

port7: f64

Port 7 power [Amp]

port8: f64

Port 8 power [Amp]

Methods

impl RNHP

pub fn from_raw(raw: RNHP_raw) -> RNHP

Convert a RNHP_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})