Struct psas_telemetry::messages::seqe::SEQE

pub struct SEQE {
    pub port: f64,
    pub expected: f64,
    pub received: f64,
}

Error message from the flight computer if it receives and out-of-order packet

Fields
port: f64

Which port the packet was recieved on []

expected: f64

Expected sequence number []

received: f64

Actual sequence number recieved []

Methods

impl SEQE

pub fn from_raw(raw: SEQE_raw) -> SEQE

Convert a SEQE_raw struct to a converted units struct by applying scaleing

pub fn to_csv_line(&self) -> String

Convert values to a single line suitable for a CSV file (ex: "12.0,-15.2,1352.8123,")

pub fn to_dict(&self) -> HashMap<String, f64>

convert values to a dictionary where the key is a string (ex: {'value_1': 12.0, 'value_2': -15.2})