Struct psas_telemetry::messages::Header

pub struct Header {
    pub fourcc: [u8; 4],
    pub timestamp: u64,
    pub size: usize,
}

A message header

Fields
fourcc: [u8; 4]

The four character code, determines the type of message

timestamp: u64

Time received in nanoseconds from message epoch

size: usize

The size of the upcoming message body, in bytes