Struct psas_telemetry::Stats

pub struct Stats {
    pub n_seqn: u64,
    pub first_seqn: u32,
    pub last_seqn: u32,
    pub n_bytes: u64,
    pub n_messages: u64,
    pub message_counts: HashMap<[u8; 4], u64>,
}

Statistics about a set of telemetry messages

Fields
n_seqn: u64

Number of SEQN (sequence number) encountered

first_seqn: u32

The value of the first SEQN (sequence number) encountered

last_seqn: u32

The value of the last SEQN (sequence number) encountered

n_bytes: u64

Total number or bytes successfully read including unknown messages

n_messages: u64

Total number of messages encountered

message_counts: HashMap<[u8; 4], u64>

A map of the number of messages encountered, keyed by their four character code